shopai-backend/requirements.txt
weaviate-client==v4.4b2
fastapi
openai
requests
uvicorn
python-dotenv
SQLAlchemy
psycopg2
google-cloud-aiplatform